load_tests/requirements.txt
locust==2.15.1
pyjwt==2.6.0
requests==2.28.2
python-dateutil==2.8.2